refactor(tooltip): stories
