--- ui-parts/tooltip/stories/interfaces.ts
@@ -1,7 +1,0 @@
-import type { TooltipProps } from '@atls-ui-parts/tooltip'
-
-export interface StoryTooltipProps
-  extends Pick<TooltipProps, 'animated' | 'arrow' | 'offset' | 'placement' | 'trigger'> {
-  customContainer: boolean
-  styledContainer: boolean
-}
