+++ ui-parts/tooltip/stories/tooltip.stories.tsx
@@ -1,9 +1,49 @@
-import type { Meta }              from '@storybook/react'
-import type { StoryObj }          from '@storybook/react'
+/* eslint-disable react/jsx-no-leaked-render */
 
-import type { StoryTooltipProps } from './interfaces.js'
+import type { TooltipProps }              from '@atls-ui-parts/tooltip'
+import type { Meta }                      from '@storybook/react'
+import type { StoryObj }                  from '@storybook/react'
+import type { ReactElement }              from 'react'
 
-import { StoryTooltip }           from './story-tooltip.js'
+import { Tooltip }                        from '@atls-ui-parts/tooltip'
+
+import { storyContainerAppearanceStyles } from './styles.css.js'
+import { storyContainerShapeStyles }      from './styles.css.js'
+import { storyTriggerStyles }             from './styles.css.js'
+
+interface StoryTooltipProps
+  extends Pick<TooltipProps, 'animated' | 'arrow' | 'offset' | 'placement' | 'trigger'> {
+  customContainer: boolean
+  styledContainer: boolean
+}
+
+const storyContainerClassName = [storyContainerAppearanceStyles, storyContainerShapeStyles].join(
+  ' '
+)
+
+const renderTooltip = ({
+  animated,
+  arrow,
+  customContainer,
+  offset,
+  placement,
+  styledContainer,
+  trigger,
+}: StoryTooltipProps): ReactElement => (
+  <Tooltip
+    animated={animated}
+    arrow={arrow && (customContainer || styledContainer) ? { fill: 'green' } : arrow}
+    appearance={styledContainer ? { container: storyContainerAppearanceStyles } : undefined}
+    offset={offset}
+    placement={placement}
+    shape={styledContainer ? { container: storyContainerShapeStyles } : undefined}
+    trigger={trigger}
+    text='Tooltip text'
+    container={customContainer ? <div className={storyContainerClassName} /> : undefined}
+  >
+    <div className={storyTriggerStyles}>Trigger</div>
+  </Tooltip>
+)
 
 const meta: Meta<StoryTooltipProps> = {
   title: 'Components/Tooltip',
@@ -70,12 +110,12 @@ const meta: Meta<StoryTooltipProps> = {
 export default meta
 
 export const Base: StoryObj<StoryTooltipProps> = {
-  render: StoryTooltip,
+  render: renderTooltip,
 }
 
 export const StyledContainer: StoryObj<StoryTooltipProps> = {
   args: {
     styledContainer: true,
   },
-  render: StoryTooltip,
+  render: renderTooltip,
 }
